[BE] refactor: randomgenerator 패키지 generator로 경로 변경 및 참조 수정 #774
