--- backend/app/src/main/java/com/onair/hearit/app/explore/application/scorefactor/randomgenerator/RandomNumberGenerator.java
@@ -1,6 +1,0 @@
-package com.onair.hearit.app.explore.application.scorefactor.randomgenerator;
-
-public interface RandomNumberGenerator {
-
-    double getDouble();
-}
