+++ backend/app/src/test/java/com/onair/hearit/app/explore/application/ExploreScoreInitializerTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import com.onair.hearit.app.explore.application.scorefactor.BookmarkScoreFactor;
 import com.onair.hearit.app.explore.application.scorefactor.RandomScoreFactor;
 import com.onair.hearit.app.explore.application.scorefactor.RecencyScoreFactor;
-import com.onair.hearit.app.explore.application.scorefactor.randomgenerator.DefaultRandomNumberGenerator;
+import com.onair.hearit.app.explore.application.scorefactor.generator.DefaultRandomNumberGenerator;
 import com.onair.hearit.app.fixture.DbHelper;
 import com.onair.hearit.core.config.DataSourceConfig;
 import com.onair.hearit.core.domain.Category;
