+++ backend/app/src/test/java/com/onair/hearit/app/explore/application/HearitExploreServiceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.BDDMockito.given;
 import com.onair.hearit.app.explore.application.scorefactor.BookmarkScoreFactor;
 import com.onair.hearit.app.explore.application.scorefactor.RandomScoreFactor;
 import com.onair.hearit.app.explore.application.scorefactor.RecencyScoreFactor;
-import com.onair.hearit.app.explore.application.scorefactor.randomgenerator.RandomNumberGenerator;
+import com.onair.hearit.app.explore.application.scorefactor.generator.RandomNumberGenerator;
 import com.onair.hearit.app.explore.application.scoreprocessor.GuestExploreScoreProcessor;
 import com.onair.hearit.app.explore.application.scoreprocessor.MemberExploreScoreProcessor;
 import com.onair.hearit.app.explore.dto.CursorRequest;
@@ -150,7 +150,7 @@ class HearitExploreServiceTest {
         // then
         assertThat(secondResponse.content())
                 .extracting("id")
-                .doesNotContain(newHearit.getId());
+                .doesNotContain(newHearit.getId()).isEmpty();
     }
 
     @DisplayName("게스트가 탐색 요청 시, 최신, 랜덤 순으로 정렬하여 반환한다")
